# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Reserveat API
  version: 1.0.0
  description: API for reserving tables in the restaurants using Reserveat
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
paths:
...
  /reservations/tables/{tableId}:
    post:
      summary: Create a reservation for a table at the specified location
      operationId: createReservation
      tags:
        - Reservation
      parameters:
        - in: path
          name: tableId
          required: true
          description: ID of the table
          schema:
            type: integer
      requestBody:
        description: Request body for creating a reservation
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/ReservationInput.yaml'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: './components/schemas/ReservationOutput.yaml'
        '400':
          $ref: './components/responses/BadRequest.yaml'
        '401':
          $ref: './components/responses/Unauthorized.yaml'
        '404':
          $ref: './components/responses/NotFound.yaml'
        '409':
          $ref: './components/responses/ReservationConflict.yaml'
  /reservations/{reservationId}:
    get:
      summary: Get a reservation by ID
      operationId: getReservationById
      tags:
        - Reservation
      parameters:
        - name: reservationId
          in: path
          description: ID of the reservation to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components/schemas/ReservationOutput.yaml'
        '401':
          $ref: './components/responses/Unauthorized.yaml'
        '404':
          $ref: './components/responses/NotFound.yaml'
    delete:
      summary: Delete a reservation
      operationId: deleteReservation
      tags:
        - Reservation
      parameters:
        - in: path
          name: reservationId
          description: ID of the reservation to delete
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Reservation deleted successfully
        '404':
          $ref: './components/responses/NotFound.yaml'
...